Bank Marketing

The data is related with direct marketing campaigns (phone calls) of a Portuguese banking institution. The classification goal is to predict if the client will subscribe a term deposit (variable `subscribed`).

Format

a data frame with 4,521 observations of 17 variables. age (numeric) jobtype of job maritalmarital status (categorical: "married","divorced","single"; note: "divorced" means divorced or widowed) education (categorical: "unknown","secondary","primary","tertiary") defaulthas credit in default? (binary: "yes","no") balanceaverage yearly balance, in euros (numeric) housinghas housing loan? (binary: "yes","no") loanhas personal loan? (binary: "yes","no") contactcontact communication type (categorical: "unknown","telephone","cellular") daylast contact day of the month (numeric) monthlast contact month of year (categorical: "jan", "feb", "mar", ..., "nov", "dec") durationlast contact duration, in seconds (numeric) campaignnumber of contacts performed during this campaign and for this client (numeric, includes last contact) pdaysnumber of days that passed by after the client was last contacted from a previous campaign (numeric, -1 means client was not previously contacted) previousnumber of contacts performed before this campaign and for this client (numeric) poutcomeoutcome of the previous marketing campaign (categorical: "unknown","other","failure","success") subscribedhas the client subscribed a term deposit? (binary: "yes","no")

Source

https://archive.ics.uci.edu/dataset/222/bank+marketing