Data from the Diagnostic Assessment and Achievement of College Skills (DAACS)

This is data part of a larger randomized control trial designed to test the effect of DAACS on student success. These data are from the treatment group at a large online college. DAACS was embedded within orientation. Although students were expected to complete orientation there were no consequences for not doing so. As a result some or all DAACS data are missing for a large proportion of the students.

Format

A data frame with 5,154 observations of 14 variables.

retained

whether the student was retained for a second semester

srl

the student's self-regulated learning score

math

the student's mathematics score

reading

the student's reading score

writing

the student's writing score

income

the student's income level as an ordered factor

employment

the student's employment status

ell

whether the student is an English Language Learner

ed_mother

the highest education level of the student's mother

ed_father

the highest education level of the student's father

ethnicity

the ethnicity of the student

gender

the gender of the student

military

whether the student is active in the military

age

the age of the student at the time of enrollment

page_views

the number of feedback pages the student viewed

Source

https://daacs.net

Details

For more information about DAACS, see https://daacs.net.