Programme of International Student Assessment

This data is from the 2019 PISA implementation for the United States. It is used to predict whether a student attends a public or private school. The dependent variable (`Public`) is is imbalanced with approximately 93

Format

A data frame with 5,233 observations of 45 variables.

Public

Whether student attends a public or private school.

ST04Q01

Sex

ST05Q01

Attend <ISCED 0>

ST06Q01

Age at <ISCED 1>

ST07Q01

Repeat <ISCED 1>

ST08Q01

At Home - Mother

ST08Q02

At Home - Father

ST08Q03

At Home - Brothers

ST08Q04

At Home - Sisters

ST08Q05

At Home - Grandparents

ST08Q06

At Home - Others

ST10Q01

Mother <Highest Schooling>

ST12Q01

Mother Current Job Status

ST14Q01

Father <Highest Schooling>

ST16Q01

Father Current Job Status

ST19Q01

Language at home

ST20Q01

Possessions desk

ST20Q02

Possessions own room

ST20Q03

Possessions study place

ST20Q04

Possessions computer

ST20Q05

Possessions software

ST20Q06

Possessions Internet

ST20Q07

Possessions literature

ST20Q08

Possessions poetry

ST20Q09

Possessions art

ST20Q10

Possessions textbooks

ST20Q12

Possessions dictionary

ST20Q13

Possessions dishwasher

ST21Q01

How many cellular phones

ST21Q02

How many televisions

ST21Q03

How many computers

ST21Q04

How many cars

ST21Q05

How many rooms bath or shower

ST22Q01

How many books at home

ST23Q01

Reading Enjoyment Time

ST31Q01

<Enrich> in <test lang>

ST31Q02

<Enrich> in <mathematics>

ST31Q03

<Enrich> in <science>

ST31Q05

<Remedial> in <test lang>

ST31Q06

<Remedial> in <mathematics>

ST31Q07

<Remedial> in <science>

ST32Q01

Out of school lessons <test lang>

ST32Q02

Out of school lessons <maths>

ST32Q03

Out of school lessons <science>

Source

https://www.pisa.oecd.org

Details

This dataset was modified from the original data provided by OECD. The [`pisa`](https://github.com/jbryer/pisa) R package on Github provides the complete data for the 2019 administration.